+++ sweat_test.py
@@ -1,0 +1,7 @@
+import downloader
+
+option = {
+    'name': 'Sweat Tests',
+    'url_code': 'QEHSWT',
+    'handler': downloader.handle_download_url,
+}
