+++ vitamin_d.py
@@ -1,0 +1,7 @@
+import downloader
+
+option = {
+    'name': 'Vitamin D',
+    'url_code': 'QEHVITD',
+    'handler': downloader.handle_download_url,
+}
